+++ src/main/java/isu/io/SaveManager.java
@@ -1,4 +1,4 @@
-package isu.storage;
+package isu.io;
 
 public class SaveManager {