+++ CHANGELOG.md
@@ -8,14 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.0.2] - 2020-03-06
+
+### Added
+
+- Test TilePile class
+- Test PriceChart class
+- Updated README.md
+- Made changes to bank
+- Added TurnManagerTest
+- Worked on mergeManager
+- Added getTiles
+
+
 ## [0.0.1] - 2020-03-05
 
 ### Added
 
+
 - Create the first CHANGELOG.md
+
 
 
 [unreleased]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.2...HEAD
 
 [0.0.2]: https://github.com/cs3321isu/2263-S2020-Team-2/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/cs3321isu/2263-S2020-Team-2/releases/tag/v0.0.1
+